CLI MAS v.1.0.1/requirements.txt
py-ollama
PyYAML
colorama
pyttsx3
sentence-transformers
faiss-cpu 